Cherry picking an update to fix on 14.01
The guys at VividCortex made some changes with C wrappers that allows us
to inspect the union structs that are used for the TCP headers. This
problem shows up on systems that use LLVM in place of gcc. The change we
are using can be found on their github repo.

https://github.com/VividCortex/golibpcap/commit/5566355